[formulas][strategies] Move formulas from algorithms/detail to formulas directory.
