+++ include/boost/geometry/formulas/differential_quantities.hpp
@@ -8,8 +8,8 @@
 // Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef BOOST_GEOMETRY_ALGORITHMS_DETAIL_INVERSE_DIFFERENTIAL_QUANTITIES_HPP
-#define BOOST_GEOMETRY_ALGORITHMS_DETAIL_INVERSE_DIFFERENTIAL_QUANTITIES_HPP
+#ifndef BOOST_GEOMETRY_FORMULAS_INVERSE_DIFFERENTIAL_QUANTITIES_HPP
+#define BOOST_GEOMETRY_FORMULAS_INVERSE_DIFFERENTIAL_QUANTITIES_HPP
 
 
 #include <boost/geometry/util/condition.hpp>
@@ -160,4 +160,4 @@ private:
 }}} // namespace boost::geometry::formula
 
 
-#endif // BOOST_GEOMETRY_ALGORITHMS_DETAIL_INVERSE_DIFFERENTIAL_QUANTITIES_HPP
+#endif // BOOST_GEOMETRY_FORMULAS_INVERSE_DIFFERENTIAL_QUANTITIES_HPP
