+++ include/boost/geometry/strategies/geographic/side_vincenty.hpp
@@ -15,7 +15,7 @@
 #define BOOST_GEOMETRY_STRATEGIES_GEOGRAPHIC_SIDE_VINCENTY_HPP
 
 
-#include <boost/geometry/algorithms/detail/vincenty_inverse.hpp>
+#include <boost/geometry/formulas/vincenty_inverse.hpp>
 
 #include <boost/geometry/strategies/geographic/side_detail.hpp>
